Copy images before running qemu-img convert
After a recent update to libvirt, qemu-img is trying to attain a shared
"write" lock on images when doing a convert/info. We cannot do this
since the backing image is read-only and in use by the running VM.
Instead we copy the file first and then run qemu-img convert.